backend/requirements.txt
Flask
pdf2image
PyPDF2
openai
pandas
werkzeug
pdfminer
pdfplumber
pdftotext
python-dotenv